+++ src/libcsg/modules/io/h5mdtrajectoryreader.cc
@@ -23,6 +23,8 @@
 // Third party includes
 #include <hdf5.h>
 
+#include <boost/regex.hpp>
+
 // Local private VOTCA includes
 #include "h5mdtrajectoryreader.h"
 
@@ -67,10 +69,6 @@ bool H5MDTrajectoryReader::Open(const std::string &file) {
     throw ios_base::failure("Wrong version of H5MD file.");
   }
 
-  // Clean up.
-  H5Aclose(at_version);
-  H5Gclose(g_h5md);
-
   // Checks if the particles group exists and what is the number of members.
   particle_group_ = H5Gopen(file_id_, "particles", H5P_DEFAULT);
   CheckError(particle_group_, "Unable to open /particles group.");
@@ -80,10 +78,28 @@ bool H5MDTrajectoryReader::Open(const std::string &file) {
     throw ios_base::failure("The particles group is empty.");
   }
 
+  // Check if the unit module is enabled.
+  hid_t modules = H5Gopen(g_h5md, "modules", H5P_DEFAULT);
+  std::cout << "open modules group" << modules << std::endl;
+  if (modules > 0) {
+    hid_t module_units = H5Gopen(modules, "units", H5P_DEFAULT);
+    if (module_units > 0) {
+      std::cout << "found units module - position and forces will be scaled "
+                << std::endl;
+      unit_module_enabled_ = true;
+      H5Gclose(module_units);
+    }
+    H5Gclose(modules);
+  }
+
   first_frame_ = true;
 
   // Handle errors by internal check up.
   H5Eset_auto2(H5E_DEFAULT, nullptr, nullptr);
+
+  // Clean up.
+  H5Aclose(at_version);
+  H5Gclose(g_h5md);
 
   return true;
 }
@@ -107,7 +123,6 @@ void H5MDTrajectoryReader::Initialize(Topology &top) {
       H5Gopen(particle_group_, position_group_name.c_str(), H5P_DEFAULT);
   CheckError(atom_position_group_,
              "Unable to open " + position_group_name + " group");
-
   idx_frame_ = -1;
   ds_atom_position_ = H5Dopen(atom_position_group_, "value", H5P_DEFAULT);
   CheckError(ds_atom_position_,
@@ -144,9 +159,9 @@ void H5MDTrajectoryReader::Initialize(Topology &top) {
          << endl;
     // Sets box size.
     m = Eigen::Matrix3d::Zero();
-    m(0, 0) = box[0];
-    m(1, 1) = box[1];
-    m(2, 2) = box[2];
+    m(0, 0) = box[0] * length_scaling_;
+    m(1, 1) = box[1] * length_scaling_;
+    m(2, 2) = box[2] * length_scaling_;
     top.setBox(m);
     has_box_ = H5MDTrajectoryReader::STATIC;
   }
@@ -186,6 +201,13 @@ void H5MDTrajectoryReader::Initialize(Topology &top) {
     cout << "H5MD: has /id group" << endl;
   } else {
     has_id_group_ = H5MDTrajectoryReader::NONE;
+  }
+
+  // Reads unit system - if enabled.
+  if (unit_module_enabled_) {
+    length_scaling_ = ReadScaleFactor(ds_atom_position_, "position");
+    force_scaling_ = ReadScaleFactor(ds_atom_force_, "force");
+    velocity_scaling_ = ReadScaleFactor(ds_atom_velocity_, "velocity");
   }
 
   // Gets number of particles and dimensions.
@@ -235,9 +257,9 @@ bool H5MDTrajectoryReader::NextFrame(Topology &top) {  // NOLINT const reference
     std::unique_ptr<double[]> box = std::unique_ptr<double[]>{new double[3]};
     ReadBox(ds_edges_group_, H5T_NATIVE_DOUBLE, idx_frame_, box);
     m = Eigen::Matrix3d::Zero();
-    m(0, 0) = box.get()[0];
-    m(1, 1) = box.get()[1];
-    m(2, 2) = box.get()[2];
+    m(0, 0) = box.get()[0] * length_scaling_;
+    m(1, 1) = box.get()[1] * length_scaling_;
+    m(2, 2) = box.get()[2] * length_scaling_;
     cout << "Time dependent box:" << endl;
     cout << m << endl;
   }
@@ -273,9 +295,9 @@ bool H5MDTrajectoryReader::NextFrame(Topology &top) {  // NOLINT const reference
   for (Index at_idx = 0; at_idx < N_particles_; at_idx++) {
     double x, y, z;
     Index array_index = at_idx * vec_components_;
-    x = positions[array_index];
-    y = positions[array_index + 1];
-    z = positions[array_index + 2];
+    x = positions[array_index] * length_scaling_;
+    y = positions[array_index + 1] * length_scaling_;
+    z = positions[array_index + 2] * length_scaling_;
     // Set atom id, or it is an index of a row in dataset or from id dataset.
     Index atom_id = at_idx;
     if (has_id_group_ != H5MDTrajectoryReader::NONE) {
@@ -296,17 +318,17 @@ bool H5MDTrajectoryReader::NextFrame(Topology &top) {  // NOLINT const reference
     b->setPos(Eigen::Vector3d(x, y, z));
     if (has_velocity_ == H5MDTrajectoryReader::TIMEDEPENDENT) {
       double vx, vy, vz;
-      vx = velocities[array_index];
-      vy = velocities[array_index + 1];
-      vz = velocities[array_index + 2];
+      vx = velocities[array_index] * velocity_scaling_;
+      vy = velocities[array_index + 1] * velocity_scaling_;
+      vz = velocities[array_index + 2] * velocity_scaling_;
       b->setVel(Eigen::Vector3d(vx, vy, vz));
     }
 
     if (has_force_ == H5MDTrajectoryReader::TIMEDEPENDENT) {
       double fx, fy, fz;
-      fx = forces[array_index];
-      fy = forces[array_index + 1];
-      fz = forces[array_index + 2];
+      fx = forces[array_index] * force_scaling_;
+      fy = forces[array_index + 1] * force_scaling_;
+      fz = forces[array_index + 2] * force_scaling_;
       b->setF(Eigen::Vector3d(fx, fy, fz));
     }
   }
@@ -343,6 +365,47 @@ void H5MDTrajectoryReader::ReadBox(hid_t ds, hid_t ds_data_type, Index row,
     throw std::runtime_error("Error ReadScalarData: " +
                              boost::lexical_cast<std::string>(status));
   }
+}
+
+double H5MDTrajectoryReader::ReadScaleFactor(const hid_t &ds,
+                                             const std::string &unit_type) {
+  hid_t unit_attr = H5Aopen(ds, "unit", H5P_DEFAULT);
+  double scaling_factor = 1.0;
+  if (unit_attr > 0) {
+    hid_t type_id = H5Aget_type(unit_attr);
+    hid_t atype_mem = H5Tget_native_type(type_id, H5T_DIR_ASCEND);
+    hsize_t str_size = H5Tget_size(type_id);
+    char buffer[80] = {0};  // buffer for attribute
+    H5Aread(unit_attr, atype_mem, &buffer);
+    H5Tclose(atype_mem);
+    H5Tclose(type_id);
+    std::string value = std::string(buffer, str_size);
+
+    // Read units
+    tools::Tokenizer tok(value, " ");
+    for (auto v : tok) {
+      boost::smatch suffix_match;
+      int unit_pow = 1;
+      if (boost::regex_match(v, suffix_match, suffix_units,
+                             boost::match_extra)) {
+        // If the prefix has numeric suffix then use it in the formula for
+        // scaling factor.
+        unit_pow = std::stoi(suffix_match[2]);
+      }
+      auto votca_scaling_factor = votca_units_scaling_factors.find(v);
+      if (votca_scaling_factor != votca_units_scaling_factors.end()) {
+        scaling_factor =
+            scaling_factor * pow(votca_scaling_factor->second, unit_pow);
+      }
+    }
+    std::cout << "Found units " << value << " for " << unit_type << ".";
+    if (scaling_factor != 1.0) {
+      std::cout << " Values scaled by a factor " << scaling_factor;
+    }
+    std::cout << std::endl;
+  }
+
+  return scaling_factor;
 }
 
 }  // namespace csg